Aplikacja webowa portalu filatelistycznego/projekt-uaim-web/backend/app/requirements.txt
flask
flask_sqlalchemy
flask_cors
psycopg2-binary
mysqlclient
flask-bcrypt
flask-jwt-extended 
datetime
pillow
requests